+++ ApocalypticFastFood/ApocalypticFastFood/Adapters/Repositories/DatabaseRepositoryAdapter.cs
@@ -4,12 +4,12 @@ public class DatabaseRepositoryAdapter : IOrderRepository
 {
     private readonly DatabaseService _db;
     public DatabaseRepositoryAdapter(DatabaseService db) => _db = db;
-    public void SaveOrder(int orderId, double total, double discount) => _db.SaveOrder(orderId, total, discount);
+    public void SaveOrder(int orderId, decimal total, decimal discount) => _db.SaveOrder(orderId, total, discount);
 }
 
 public class DatabaseService
 {
-    public void SaveOrder(int orderId, double total, double disc)
+    public void SaveOrder(int orderId, decimal total, decimal disc)
     {
         Console.WriteLine("Saving order " + orderId + " to database");
     }
